requirements.txt
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
blas=1.0=mkl
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.12.12=h06a4308_0
certifi=2021.5.30=py36h06a4308_0
cudatoolkit=11.0.221=h6bb024c_0
cycler=0.11.0=pypi_0
dataclasses=0.8=pyh4f3eec9_6
decorator=4.4.2=pypi_0
ffmpeg=4.3=hf484d3e_0
freetype=2.12.1=h4a9f257_0
giflib=5.2.1=h5eee18b_3
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
imageio=2.15.0=pypi_0
importlib-resources=5.4.0=pypi_0
install=1.3.5=pypi_0
intel-openmp=2022.1.0=h9e868ea_3769
jpeg=9b=h024ee3a_2
kiwisolver=1.3.1=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.3=he6710b0_2
libgcc-ng=13.2.0=h807b86a_3
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=13.2.0=h807b86a_3
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=13.2.0=h7e041cc_3
libtasn1=4.19.0=h5eee18b_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuv=1.44.2=h5eee18b_0
libwebp=1.2.0=h89dd481_0
libwebp-base=1.2.0=h27cfd23_0
lz4-c=1.9.4=h6a678d5_0
matplotlib=3.3.4=pypi_0
mkl=2022.1.0=hc2b9512_224
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=2.5.1=pypi_0
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
numpy=1.19.5=pypi_0
olefile=0.46=pyhd3eb1b0_0
opencv-python=4.3.0.36=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1w=h7f8727e_0
pandas=1.1.0=pypi_0
pillow=8.4.0=pypi_0
pip=21.3.1=pypi_0
pyparsing=3.1.1=pypi_0
python=3.6.13=h12debd9_1
python-dateutil=2.8.2=pypi_0
python_abi=3.6=2_cp36m
pytorch=1.7.1=py3.6_cuda11.0.221_cudnn8.0.5_0
pytorch-mutex=1.0=cuda
pytz=2023.3.post1=pypi_0
pywavelets=1.1.1=pypi_0
readline=8.2=h5eee18b_0
scikit-image=0.17.2=pypi_0
scipy=1.5.4=pypi_0
setuptools=59.6.0=pypi_0
six=1.16.0=pypi_0
sqlite=3.41.2=h5eee18b_0
tifffile=2020.9.3=pypi_0
timm=0.5.4=pypi_0
tk=8.6.12=h1ccaba5_0
torchaudio=0.7.2=py36
torchvision=0.8.2=py36_cu110
tqdm=4.64.1=pypi_0
typing_extensions=4.1.1=pyh06a4308_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.4.5=h5eee18b_0
zipp=3.6.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.4.9=haebb681_0
